# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ForumDesk Web API",
    "version": "1.0.0",
    "description": "Web visitor conversations bridged to Telegram Forum Topics."
  },
  "servers": [
    {
      "url": "https://support.example.com"
    }
  ],
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Healthy"
          }
        }
      }
    },
    "/api/v1/conversations": {
      "post": {
        "summary": "Create a web conversation",
        "security": [
          {
            "integrationBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conversation created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateConversationResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/v1/conversations/{conversation_id}/messages": {
      "parameters": [
        {
          "name": "conversation_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "Send a visitor message",
        "security": [
          {
            "visitorBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "text"
                ],
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 4000
                  },
                  "reply_to_message_id": {
                    "type": "string",
                    "description": "Optional visible message ID to quote"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Message"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      },
      "get": {
        "summary": "Poll visible conversation messages",
        "security": [
          {
            "visitorBearer": []
          }
        ],
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message page",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Message"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "next_cursor": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "integrationBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "WEB_API_KEY; server-side only"
      },
      "visitorBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Conversation-scoped visitor_token"
      }
    },
    "schemas": {
      "CreateConversation": {
        "type": "object",
        "required": [
          "visitor_name"
        ],
        "additionalProperties": false,
        "properties": {
          "visitor_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "email": {
            "type": "string",
            "maxLength": 254
          },
          "external_id": {
            "type": "string",
            "maxLength": 128
          },
          "page_url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "CreateConversationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "visitor_token": {
                "type": "string",
                "description": "Returned once"
              },
              "status": {
                "type": "string",
                "const": "open"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "conversation_id": {
            "type": "string"
          },
          "sequence": {
            "type": "integer"
          },
          "direction": {
            "type": "string",
            "enum": [
              "customer",
              "staff",
              "system"
            ]
          },
          "event": {
            "type": "string",
            "enum": [
              "message_hidden"
            ]
          },
          "target_message_id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "staff_name": {
            "type": "string"
          },
          "visible": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "reply_to_message_id": {
            "type": "string",
            "description": "Quoted message ID"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Invalid input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}